backend/requirements.txt
fastapi==0.110.1
uvicorn==0.27.1
pillow==10.3.0
pytesseract==0.3.10
easyocr==1.7.1
opencv-python==4.9.0.80
torch==2.2.2
torchvision==0.17.2
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
matplotlib==3.8.4
python-multipart==0.0.9
